backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
websockets==13.1
google-genai==1.65.0
python-dotenv==1.0.1
pydantic-settings==2.4.0
